+++ packages/world-modules/test/CallWithSignatureModule.t.sol
@@ -19,11 +19,11 @@ import { REGISTRATION_SYSTEM_ID } from "@latticexyz/world/src/modules/init/const
 import { createWorld } from "@latticexyz/world/test/createWorld.sol";
 import { WorldTestSystem } from "@latticexyz/world/test/World.t.sol";
 
-import { Unstable_CallWithSignatureModule } from "../src/modules/delegation/Unstable_CallWithSignatureModule.sol";
-import { Unstable_CallWithSignatureSystem } from "../src/modules/delegation/Unstable_CallWithSignatureSystem.sol";
-import { IUnstable_CallWithSignatureErrors } from "../src/modules/delegation/IUnstable_CallWithSignatureErrors.sol";
-import { getSignedMessageHash } from "../src/modules/delegation/getSignedMessageHash.sol";
-import { ECDSA } from "../src/modules/delegation/ECDSA.sol";
+import { Unstable_CallWithSignatureModule } from "../src/modules/callwithsignature/Unstable_CallWithSignatureModule.sol";
+import { Unstable_CallWithSignatureSystem } from "../src/modules/callwithsignature/Unstable_CallWithSignatureSystem.sol";
+import { IUnstable_CallWithSignatureErrors } from "../src/modules/callwithsignature/IUnstable_CallWithSignatureErrors.sol";
+import { getSignedMessageHash } from "../src/modules/callwithsignature/getSignedMessageHash.sol";
+import { ECDSA } from "../src/modules/callwithsignature/ECDSA.sol";
 
 contract Unstable_CallWithSignatureModuleTest is Test, GasReporter {
   using WorldResourceIdInstance for ResourceId;