+++ packages/common/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 2.0.5
+
+### Patch Changes
+
+- a9e8a407: Fixed `getNonceManager` to correctly pass all options to `createNonceManager`.
+  - @latticexyz/schema-type@2.0.5
+
 ## 2.0.4
 
 ### Patch Changes